+++ Sources/GestureButton/GestureButtonTimer.swift
@@ -8,8 +8,7 @@
 
 import Foundation
 
-/// This internal class can be used to repeat an action when
-/// a button is kept pressed.
+/// This class can be used to repeat an action when a button is kept pressed.
 public class GestureButtonTimer: ObservableObject {
 
     /// Create a custom gesture button timer.